(PC-33870) feat(reaction): reintegrate cookie check to reaction modal
